Add files via upload
In this version I have attempted to produce a probabilistic system, storing all of the cross junction in a list of dictionaries. then incrementing the road usage by 1 every time a node travels on it. It then works out the probability that the road will be used using the summation of road usages in the list and then dividing the i uses by the sum. There is another function that then gets the branches of each cross junction using a filter and then gets works out if the road is vacant or not. If it is then it uses that as a factor to determine the lowest costing route.